src/main/java/com/aaron/challenge/flightadvisor/comments/CommentService.java
package com.aaron.challenge.flightadvisor.comments;

import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Example;
import org.springframework.data.domain.ExampleMatcher;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.Optional;

@RequiredArgsConstructor

@Service
public class CommentService {

    public static final ExampleMatcher matcher = ExampleMatcher
            .matching()
            .withIgnorePaths("id")
            .withIgnoreNullValues()
            .withIgnoreCase(true)
            .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING);

    private final CommentRepository commentRepository;

    public Optional<Comment> findById(String id) {
        return commentRepository.findById(id);
    }

    public Comment create(Comment comment) {
        return commentRepository.save(comment);
    }

    public Comment update(Comment comment) {
        return commentRepository.save(comment);
    }

    public Page<Comment> search(Comment comment, Pageable pageable) {
        return commentRepository.findAll(Example.of(comment, matcher), pageable);
    }

    public Page<Comment> findAll(Pageable pageable) {
        return commentRepository.findAll(pageable);
    }

    public void delete(Comment comment) {
        commentRepository.delete(comment);
    }
}
